Decouple ImageSharp/ImageSharp.Web implementations
