+++ src/Umbraco.Core/Media/NoopImageDimensionExtractor.cs
@@ -1,0 +1,9 @@
+using System.Drawing;
+
+namespace Umbraco.Cms.Core.Media
+{
+    public sealed class NoopImageDimensionExtractor : IImageDimensionExtractor
+    {
+        public Size? GetDimensions(Stream? stream) => null;
+    }
+}
