+++ src/Umbraco.Web.Common/Umbraco.Web.Common.csproj
@@ -18,7 +18,6 @@
     <PackageReference Include="Microsoft.AspNetCore.Mvc.Razor.RuntimeCompilation" Version="7.0.0-preview.7.22376.6" />
     <PackageReference Include="MiniProfiler.AspNetCore.Mvc" Version="4.2.22" />
     <PackageReference Include="Serilog.AspNetCore" Version="6.0.1" />
-    <PackageReference Include="SixLabors.ImageSharp.Web" Version="2.0.2" />
     <PackageReference Include="Smidge.InMemory" Version="4.1.1" />
     <PackageReference Include="Smidge.Nuglify" Version="4.1.1" />
   </ItemGroup>
