+++ tests/Umbraco.Tests.UnitTests/Umbraco.Web.Common/ImageProcessors/CropWebProcessorTests.cs
@@ -11,7 +11,7 @@ using SixLabors.ImageSharp.Web;
 using SixLabors.ImageSharp.Web.Commands;
 using SixLabors.ImageSharp.Web.Commands.Converters;
 using SixLabors.ImageSharp.Web.Middleware;
-using Umbraco.Cms.Web.Common.ImageProcessors;
+using Umbraco.Cms.Imaging.ImageSharp.ImageProcessors;
 
 namespace Umbraco.Cms.Tests.UnitTests.Umbraco.Web.Common.ImageProcessors;
